revert(refactor): Remove fully-enabled/legacy feature flags (funnel tests)
