+++ posthog/hogql_queries/insights/funnels/test/test_funnel_correlation_actors.py
@@ -74,7 +74,8 @@ def get_actors(
     return response.results
 
 
-class TestFunnelCorrelationActors(ClickhouseTestMixin, APIBaseTest):
+class BaseTestFunnelCorrelationActors(ClickhouseTestMixin, APIBaseTest):
+    __test__ = False
     maxDiff = None
 
     def _setup_basic_test(self):
@@ -643,3 +644,7 @@ class TestFunnelCorrelationActors(ClickhouseTestMixin, APIBaseTest):
                 }
             ],
         )
+
+
+class TestFunnelCorrelationActors(BaseTestFunnelCorrelationActors):
+    __test__ = True
