+++ posthog/hogql_queries/insights/funnels/test/test_funnel_trends_actors.py
@@ -25,7 +25,9 @@ filters = {
 
 
 @freeze_time("2021-05-01")
-class TestFunnelTrendsActors(ClickhouseTestMixin, APIBaseTest):
+class BaseTestFunnelTrendsActors(ClickhouseTestMixin, APIBaseTest):
+    __test__ = False
+
     @snapshot_clickhouse_queries
     def test_funnel_trend_persons_returns_recordings(self):
         persons = journeys_for(
@@ -163,3 +165,7 @@ class TestFunnelTrendsActors(ClickhouseTestMixin, APIBaseTest):
             [next(iter(results[0][2]))["session_id"]],
             ["s1a"],
         )
+
+
+class TestFunnelTrendsActors(BaseTestFunnelTrendsActors):
+    __test__ = True
