+++ posthog/hogql_queries/insights/funnels/test/test_funnel_udf.py
@@ -3,6 +3,7 @@ from typing import cast
 
 from freezegun import freeze_time
 from posthog.test.base import ClickhouseTestMixin, _create_event, _create_person
+from unittest.mock import Mock, patch
 
 from test_funnel import PseudoFunnelActors, funnel_test_factory
 
@@ -36,6 +37,10 @@ def _create_action(**kwargs):
     return action
 
 
+use_udf_funnel_flag_side_effect = lambda key, *args, **kwargs: key == "insight-funnels-use-udf"
+
+
+@patch("posthoganalytics.feature_enabled", new=Mock(side_effect=use_udf_funnel_flag_side_effect))
 class TestFunnelBreakdownUDF(
     ClickhouseTestMixin,
     funnel_breakdown_test_factory(  # type: ignore
@@ -49,6 +54,7 @@ class TestFunnelBreakdownUDF(
     pass
 
 
+@patch("posthoganalytics.feature_enabled", new=Mock(side_effect=use_udf_funnel_flag_side_effect))
 class TestFunnelGroupBreakdownUDF(
     ClickhouseTestMixin,
     funnel_breakdown_group_test_factory(  # type: ignore
@@ -59,8 +65,9 @@ class TestFunnelGroupBreakdownUDF(
     pass
 
 
+@patch("posthoganalytics.feature_enabled", new=Mock(side_effect=use_udf_funnel_flag_side_effect))
 class TestFOSSFunnelUDF(funnel_test_factory(_create_event, _create_person)):  # type: ignore
-    def test_assert_udf_flag(self):
+    def test_assert_flag_is_on(self):
         filters = {
             "insight": INSIGHT_FUNNELS,
             "funnel_viz_type": "steps",
@@ -317,6 +324,7 @@ class TestFOSSFunnelUDF(funnel_test_factory(_create_event, _create_person)):  # 
     maxDiff = None
 
 
+@patch("posthoganalytics.feature_enabled", new=Mock(side_effect=use_udf_funnel_flag_side_effect))
 class TestFunnelConversionTimeUDF(
     ClickhouseTestMixin,
     funnel_conversion_time_test_factory(FunnelOrderType.ORDERED, PseudoFunnelActors),  # type: ignore
